Optimize variable 'verbose',
According to the assignment description, additional information should be displayed if the EC2_BACKUP_VERBOSE environment variable is set. The description does not clarify what content should the EC2_BACKUP_VERBOSE should be set, so we should display additional information when the variable EC2_BACKUP_VERBOSE exists, no matter what the content is in it.